+++ gatling-charts/src/main/scala/com/excilys/ebi/gatling/charts/report/AllSessionsReportGenerator.scala
@@ -24,7 +24,7 @@ import com.excilys.ebi.gatling.core.result.reader.DataReader
 class AllSessionsReportGenerator(runOn: String, dataReader: DataReader, componentLibrary: ComponentLibrary) extends ReportGenerator(runOn, dataReader, componentLibrary) {
 
 	def generate {
-		val series = new Series[Long, Int]("All Sessions", dataReader.numberOfActiveSessionsPerSecond(), List(ORANGE))
+		val series = new Series[Long, Long]("All Sessions", dataReader.numberOfActiveSessionsPerSecond(), List(ORANGE))
 
 		val javascript = componentLibrary.getAllSessionsJs(series)
